remove deprecated test helper and replace with jest test helper
