+++ test/redux/groups.test.js
@@ -1,5 +1,5 @@
 import deepFreeze from "deep-freeze"
-import { setupMockRetroChannel } from "../support/js/jest_test_helper"
+import { setupMockRetroChannel } from "../support/js/test_helper"
 
 import {
   reducer as groupsReducer,
